### 2020-05-31 #2280 @whedon
[PRE REVIEW]: The Basic Model Interface 2.0: A standard interface for coupling numerical models in the geosciences
**Submitting author:** @mdpiper (<a href="http://orcid.org/0000-0001-6418-277X">MARK PIPER</a>)
**Repository:** <a href="https://github.com/csdms/bmi" target ="_blank">https://github.com/csdms/bmi</a>
**Version:** v2.0
