### 2020-06-01 #2280 @danielskatz
Note to editors: this was discussed in https://github.com/openjournals/joss/issues/690 before submission
