### 2020-06-01 #2280 @danielskatz
👋 @mdpiper - Thanks for your submission to JOSS. As described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Since reopening JOSS about a little more than a week ago, we've had > 60 papers submitted and as such, yours has been put in our backlog that we will be working through over the coming weeks and months. We also are beginning to bring in new editors to move this process along.

Thanks in advance for your patience!
