### 2020-06-06 #2280 @kyleniemeyer
@jedbrown oh, sorry. I missed that. So, I think we have faced this sort of conflict before, and I'm not as concerned about this for an editor, as long as the reviewers do not have similar conflicts. @joss-eics do you agree?
