### 2020-06-06 #2280 @diehlpk
I can take this, since I have experience in coupling methods for computational fracture mechanics.
