### 2020-06-07 #2280 @diehlpk
Hi @mdpiper, I am the editor for this submission. I briefly read the paper and have following remarks:

1. There are no references in the Reference section
2. All Bib Tex entries are not resolved

Please have look into these issues.
