### 2020-06-07 d6276ce
Fix citations
I had incorrectly interpreted the directions in the docs.
