### 2020-06-08 #2280 @teuben
funny you say.   If I do accept (i might, although this is not my field, and this would be my first online review), I ran into this article today, which it now on my reading shelf:

https://abompard.wordpress.com/2020/06/07/reviews-are-hard/
