### 2020-06-08 #2280 @teuben
@mdpiper when i read the abstract, the amuse project in astropysics came to mind. have you heard of this before?   https://ascl.net/1107.007 has a reference. Comining legacy (and new) codes via a python hub.
