### 2020-06-08 #2280 @mdpiper
@teuben I hadn't seen AMUSE before, but from the examples, it looks similar to what we in the earth surface processes community are doing with [pymt](https://pymt.readthedocs.io/en/latest/). Neat! Thanks for the reference!
