### 2020-06-08 #2280 @teuben
and there is another group (in amsterdam?) that has cloned amuse for their field (hydrology?). i remember looking at their link, but they are not as active as amuse it (which still is very active).
