### 2020-06-09 #2317 @mdpiper
A quick note: in a pre-submission [issue](https://github.com/openjournals/joss/issues/690), editors suggested this submission be reviewed in the context of the language specifications and examples derived from the interface defined in the main repository.
