### 2020-06-11 #2317 @diehlpk
@mdpiper Are these listed in the JOSS paper as well? If, not would it be possible to add them?
