### 2020-06-11 #2317 @mdpiper
@diehlpk Yup! See Table 1 in the paper.
