### 2020-06-16 #2317 @diehlpk
Hi @yangbai90 thanks for your effort supporting JOSS :)
