### 2020-06-29 csdms/bmi#68 @teuben
SIDL reminders in generated files?
If certain files (e.g. bmi-c/bmi.h) are generated via the bmi.sidl and some procedure, this fact should be a comment in those files, as to yet again remind authors of the proper workflow.

https://github.com/openjournals/joss-reviews/issues/2317
