### 2020-06-29 csdms/bmi#69 @teuben
package bmic not found (in a non-standard prefix install)
I first installed bmic:

mkdir _build && cd _build
cmake .. -DCMAKE_INSTALL_PREFIX=/tmp/opt
make install

and then the example-C in the same fashion:

mkdir _build && cd _build
cmake .. -DCMAKE_INSTALL_PREFIX=/tmp/opt

when I get "No package 'bmic' found"

Is there a flag in cmake one gives to say that bmic is installed in a non-standard place, or did I not follow all install instructions.

https://github.com/openjournals/joss-reviews/issues/2317
