### 2020-06-29 csdms/bmi#69 @mdpiper
Hmm, pkg-config should locate the library. I'll disable it to see if I can reproduce this result.

(There may be some delay since I'm on vacation this week.)
