### 2020-06-30 csdms/bmi#70 @teuben
how to run ctest in bmi-example-fortran?
The readme file refers to ctest how to run units tests and examples, but I'm not familiar how to run this. perhaps the commandline how to run this would be useful in this readme file.
I did learn something new:   ctest is something that comes with cmake
