### 2020-06-30 csdms/bmi#69 @teuben
I should add that bmi-cxx has not too surprisingly the same issue with not finding bmicxx.

The fortran and python installed and tested fine (but see #70)
