### 2020-06-30 csdms/bmi#70 @teuben
I may have answered this myself:     "make test" or "ctest" runs in the _build directory. I was in the wrong (main) directory, had to be in _build.
