### 2020-07-01 csdms/bmi#69 @mcflugen
@teuben Thank you so much for looking at this!

To help *pkg-config* find things installed in non-standard locations, you can set the *PKG_CONFIG_PATH* environment variable. For example, the following works for me.

Install the *bmi-c* package,
```bash
$ mkdir _build && cd _build  # In bmi-c
$ cmake ../ -DCMAKE_INSTALL_PREFIX=/tmp/opt
$ make install
```
Install the *bmi-example-c* example,
```bash
$ mkdir _build && cd _build  # In bmi-example-c
$ export PKG_CONFIG_PATH=/tmp/opt/lib/pkgconfig
$ cmake ../ -DCMAKE_INSTALL_PREFIX=/tmp/opt
$ make install
```
Run the example,
```bash
$ /tmp/opt/bin/run_bmiheatc testing/config.txt  # In bmi-example-c
$ ls bmiheatc.out
bmiheatc.out
```
The first ten line of the output file,
```
Configuration file = testing/config.txt
Initializing... done
The 2D Heat Equation
variable = plate_surface__temperature
rank = 2
shape = 5 x 6

Time = 0.000000
0.00e+00 0.00e+00 0.00e+00 0.00e+00 0.00e+00 0.00e+00
0.00e+00 2.24e+00 2.24e+00 5.43e+00 -1.46e+00 0.00e+00
```
