### 2020-07-01 csdms/bmi#70 @mcflugen
@teuben Yes, that's correct, *ctest* needs to be run in the *_build* directory. I made the same mistake when I recently went over this. We should update the docs to clarify.
