### 2020-07-06 csdms/bmi#68 @mdpiper
@teuben Currently, the language files (like bmi-c/bmi.h) are created manually. However, it's a really good idea to generate them programmatically. @mcflugen and I are scoping out the work for this now.
