### 2020-07-06 csdms/bmi#69 @mdpiper
Hi @teuben -- in the PRs listed above, I've updated the READMEs in the bmi-example-c and bmi-example-cxx repos to try to clarify how to install to a nonstandard location. Please let me know what you think. If they address this issue satisfactorily, I'll close the issue and merge the PRs.
