### 2020-07-06 csdms/bmi#70 @mdpiper
Hi @teuben -- in the PRs listed above, I've updated the READMEs in the bmi-example-fortran, bmi-example-c, and bmi-example-cxx repos to try to clarify how to run ctest. Please let me know what you think. If they address this issue satisfactorily, I'll close the issue and merge the PRs.
