### 2020-07-10 #2317 @mdpiper
Hi @teuben, I think we've addressed your comments in https://github.com/csdms/bmi/issues/68, https://github.com/csdms/bmi/issues/69, and https://github.com/csdms/bmi/issues/70. When you have time, would you please check the open PRs listed under each issue? If they pass muster, I'll merge the PRs and close the issues.
