### 2020-07-14 csdms/bmi#70 @teuben
they all work for me now, thanks for the clarifications and README updates.
