### 2020-07-14 csdms/bmi#69 @teuben
yes, go ahead and merge. They work for me.
