### 2020-07-14 csdms/bmi#68 @teuben
given #71 this can be closed I guess.
