### 2020-07-14 #2317 @diehlpk
@mdpiper I will do the final review on the paper by the end of this week. After that

`When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (on Zenodo, figshare, or other). The authors then post the version number and archive DOI in the REVIEW issue. `
