### 2020-07-14 #2317 @diehlpk
@mdpiper

`State of the field: Do the authors describe how this software compares to other commonly-used packages?`

The reviewers agreed on this, however, I am as not as deeply in the topic would prefer if you can highlight this point in the paper a little more. So a broader audience can see this point.

Everything else is fine.
