### 2020-07-14 #2317 @mdpiper
@diehlpk Thanks for your editorial review. I'll ask some text to this point. And another thank you--this a good point to make considering model coupling is decades old in, e.g., atmospheric sciences, but still relatively new in the geosciences.
