### 2020-07-15 af91bef
Remove style formatting from table caption
