### 2020-07-20 1dfc33d
minor edits to paper.md (#73)
Just a couple of minor tweaks
