### 2020-07-21 8fe6587
Minor updates to text
