+++ paper/paper.md
@@ -96,6 +96,8 @@ and an example implementation are listed in Table 1 below.
 Detailed instructions for building the language specification and example
 are given within each repository.
 
+\  
+
 **Table 1:**
 Repositories containing BMI language specifications and examples.
 Prefix the CSDMS GitHub organization (https://github.com/csdms/) to the
