### 2020-07-21 da8230b
Broaden paper introduction (#74)
* Expand paper intro for broader audience

* Try to shift table caption after page break

* Remove style formatting from table caption

* Minor edits

* minor edits to paper.md (#73)

Just a couple of minor tweaks

* Minor updates to text

* Add line break for table header

Co-authored-by: Greg Tucker <gtucker@colorado.edu>
