+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{collins:2005,
+  title={Design and implementation of components in the Earth System Modeling Framework},
+  author={Collins, Nancy and Theurich, Gerhard and Deluca, Cecelia and Suarez, Max and Trayanov, Atanas and Balaji, Venkatramani and Li, Peggy and Yang, Weiyu and Hill, Chris and Da Silva, Arlindo},
+  journal={The International Journal of High Performance Computing Applications},
+  volume={19},
+  number={3},
+  pages={341--350},
+  year={2005},
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+}
+
 @article{epperly:2012,
   title={High-performance language interoperability for scientific
          computing through Babel},
