### 2020-07-21 #2317 @mdpiper
@diehlpk Nudging successful! The updated paper text, with a broader introduction, is in the mdpiper/paper branch. If you're OK with the changes, I'll follow your instructions upthread for finalizing the submission.
