### 2020-07-21 #2317 @diehlpk
@mdpiper The addition to the paper is sufficient and I can recommend the paper for acceptance.

Next steps are:

`
When a submission is ready to be accepted, we ask that the authors issue a new tagged release of the software (if changed), and archive it (on Zenodo, figshare, or other). The authors then post the version number and archive DOI in the REVIEW issue.`
