### 2020-07-21 #2317 @mdpiper
@diehlpk version: [v2.0](https://github.com/csdms/bmi/releases/tag/v2.0), doi: 10.5281/zenodo.3955010

On Zenodo: [Basic Model Interface 2.0](https://zenodo.org/record/3955010#.XxcnjR1MEbI)
