### 2020-07-21 #2317 @mdpiper
@diehlpk The rendered PDF is correct; however, I merged the mdpiper/paper branch into master earlier this morning. (I'm not sure if this is important, but I want to check.)
