### 2020-07-21 #2317 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
