+++ README.rst
@@ -20,6 +20,8 @@
 
    <p align="center">
 
+   <a href='https://doi.org/10.5281/zenodo.3955009'>
+     <img src='https://zenodo.org/badge/DOI/10.5281/zenodo.3955010.svg' alt='DOI'></a>
    <a href='https://bmi.readthedocs.io/en/latest/?badge=latest'>
      <img src='https://readthedocs.org/projects/bmi/badge/?version=latest' alt='Documentation Status'></a>
    <a href="https://opensource.org/licenses/MIT">
