### 2020-07-23 #2317 @Kevin-Mattheus-Moerman
@mdpiper I've reviewed your paper. It is ready for acceptance once you make the below change. You can run `@whedon generate pdf` here to update the paper once you are finished.

- [ ] Can you add countries to your affiliations?
